refactor(nimbus-provider): restructure color mode handling and improve hook organization
- Removed the deprecated ColorModeProvider component and replaced it with NimbusColorModeProvider for better clarity.
- Consolidated color mode hooks into a dedicated directory, enhancing modularity and maintainability.
- Updated imports across components to reflect the new structure, ensuring consistent usage of color mode functionality.
