+++ packages/nimbus/src/components/nimbus-provider/components/nimbus-provider.color-mode-provider.tsx
@@ -1,0 +1,8 @@
+import { ThemeProvider } from "next-themes";
+import type { ColorModeProviderProps } from "../nimbus-provider.types";
+
+export const NimbusColorModeProvider = (props: ColorModeProviderProps) => {
+  return (
+    <ThemeProvider attribute="class" disableTransitionOnChange {...props} />
+  );
+};
