+++ packages/nimbus/src/components/nimbus-provider/index.ts
@@ -1,9 +1,2 @@
+export * from "./nimbus-provider.types";
 export * from "./nimbus-provider.tsx";
-export { useColorMode, useColorModeValue } from "./color-mode";
-export type {
-  ColorModeProviderProps,
-  NimbusRouterConfig,
-  NimbusProviderProps,
-  RouterOptions,
-  TypedNimbusRouterConfig,
-} from "./nimbus-provider.types";
