+++ packages/nimbus/src/components/nimbus-provider/nimbus-provider.tsx
@@ -1,45 +1,8 @@
-import { useEffect, useState, Fragment } from "react";
 import { ChakraProvider } from "@chakra-ui/react";
-import { ColorModeProvider } from "./color-mode";
 import { I18nProvider, RouterProvider } from "react-aria";
-import { system } from "../../theme";
-import type {
-  NimbusRouterConfig,
-  NimbusProviderProps,
-  TypedNimbusRouterConfig,
-} from "./nimbus-provider.types";
-
-export function useColorScheme() {
-  const [colorScheme, setColorScheme] = useState(getCurrentColorScheme());
-
-  // Helper function to get the current color-scheme from the <html> tag
-  function getCurrentColorScheme() {
-    return (
-      document.documentElement.style.getPropertyValue("color-scheme") ||
-      localStorage.getItem("theme") ||
-      "light"
-    );
-  }
-
-  useEffect(() => {
-    const htmlElement = document.documentElement;
-
-    const observer = new MutationObserver(() => {
-      const newColorScheme = getCurrentColorScheme();
-      setColorScheme(newColorScheme);
-    });
-
-    // Observe changes to the 'style' attribute of the <html> element
-    observer.observe(htmlElement, {
-      attributes: true,
-      attributeFilter: ["style"],
-    });
-
-    return () => observer.disconnect();
-  }, []);
-
-  return colorScheme;
-}
+import { system } from "@/theme";
+import type { NimbusProviderProps } from "./nimbus-provider.types";
+import { NimbusColorModeProvider } from "./components/nimbus-provider.color-mode-provider";
 
 export function NimbusProvider({
   children,
@@ -50,9 +13,9 @@ export function NimbusProvider({
   // Inner content with all the existing providers
   const content = (
     <ChakraProvider value={system}>
-      <ColorModeProvider enableSystem={false} {...props}>
+      <NimbusColorModeProvider enableSystem={false} {...props}>
         <I18nProvider locale={locale}>{children}</I18nProvider>
-      </ColorModeProvider>
+      </NimbusColorModeProvider>
     </ChakraProvider>
   );
