+++ packages/nimbus/src/hooks/index.ts
@@ -1,2 +1,5 @@
-export { useHotkeys } from "@/hooks/use-hotkeys/use-hotkeys.ts";
-export { useCopyToClipboard } from "@/hooks/use-copy-to-clipboard/use-copy-to-clipboard.ts";
+export * from "./use-hotkeys";
+export * from "./use-copy-to-clipboard";
+export * from "./use-color-scheme";
+export * from "./use-color-mode";
+export * from "./use-color-mode-value";
