+++ packages/nimbus/src/hooks/use-color-mode-value/use-color-mode-value.ts
@@ -1,0 +1,13 @@
+import { useColorMode } from "./../use-color-mode";
+
+/**
+ * Hook that returns a value based on the current color mode.
+ *
+ * @param light - The value to return when the color mode is light
+ * @param dark - The value to return when the color mode is dark
+ * @returns The appropriate value based on the current color mode
+ */
+export function useColorModeValue<T>(light: T, dark: T) {
+  const { colorMode } = useColorMode();
+  return colorMode === "light" ? light : dark;
+}
