+++ packages/nimbus/src/hooks/use-color-mode/index.ts
@@ -1,0 +1,1 @@
+export { useColorMode } from "./use-color-mode";
