+++ packages/nimbus/src/hooks/use-color-scheme/index.ts
@@ -1,0 +1,1 @@
+export { useColorScheme } from "./use-color-scheme";
