+++ packages/nimbus/src/hooks/use-copy-to-clipboard/index.ts
@@ -1,0 +1,1 @@
+export { useCopyToClipboard } from "./use-copy-to-clipboard.ts";
