+++ packages/nimbus/src/hooks/use-hotkeys/index.ts
@@ -1,0 +1,1 @@
+export { useHotkeys } from "./use-hotkeys.ts";
